- investment route configuration
