+++ src/pages/investment/payments/index.tsx
@@ -1,0 +1,14 @@
+import React, { ReactElement } from 'react'
+
+interface Props {
+  transactions?: []
+}
+
+export default function Payments({ transactions }: Props): ReactElement {
+  console.log(transactions)
+  return (
+    <div>
+      
+    </div>
+  )
+}
